replaced 'unsigned int' with 'sample_index_t' in more places, for better 32/64 bit compatibility
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/branches/signal_views@2523 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ libkwave/Decoder.h
@@ -81,7 +81,7 @@ signals:
      * processed.
      * @param pos current position within the source, in bytes!
      */
-    void sourceProcessed(unsigned int pos);
+    void sourceProcessed(quint64 pos);
 
 protected:
