+++ plugins/codec_audiofile/AudiofileDecoder.cpp
@@ -169,7 +169,8 @@ bool AudiofileDecoder::open(QWidget *widget, QIODevice &src)
     qDebug("channels    = %d", info().tracks());
     qDebug("rate        = %0.0f", info().rate());
     qDebug("bits/sample = %d", info().bits());
-    qDebug("length      = %d samples", info().length());
+    qDebug("length      = %lu samples",
+           static_cast<unsigned long int>(info().length()));
     qDebug("format      = %d (%s)", sample_format,
                                     sample_format_name.toLocal8Bit().data());
     qDebug("-------------------------");
